Move SwimEditor component into its own module
