+++ src/ui.rs
@@ -1,7 +1,9 @@
 //! User interface using Yew, organized into components.
 
+
+pub mod assembled_view;
 pub mod console;
 pub mod hex_editor;
 pub mod regview;
+pub mod swim_editor;
 pub mod visual_datapath;
-pub mod assembled_view;
