+++ src/ui/swim_editor/mod.rs
@@ -1,0 +1,1 @@
+pub mod component;
